+++ src/types/webcast.ts
@@ -1,4 +1,5 @@
 import { Rev, RegistrationField, Admin } from '.';
+import { Video } from './video';
 import { LiteralString } from './rev';
 
 export interface Webcast {
@@ -400,16 +401,8 @@ export namespace Webcast {
         userAgent?: string;
     }
 
-    export interface Playback {
-        label: string;
-        qValue: number;
-        player: string;
-        url: string;
-        zoneId: string;
-        slideDelaySeconds: number;
-        videoFormat: string;
-        videoInstanceId: string;
-        deviceId: string;
+    export interface Playback extends Video.PlaybackUrlResult {
+
     }
 }